[REFACTOR][I] Move classes of saros.project.* into saros.filesystem
Removes VirtualFileMock as it was no longer being used.
