+++ intellij/src/saros/intellij/context/SarosIntellijContextFactory.java
@@ -13,10 +13,10 @@ import saros.filesystem.IWorkspace;
 import saros.filesystem.checksum.IChecksumCache;
 import saros.filesystem.checksum.NullChecksumCache;
 import saros.intellij.editor.EditorManager;
+import saros.intellij.filesystem.IntelliJWorkspaceImpl;
+import saros.intellij.filesystem.PathFactory;
 import saros.intellij.preferences.IntelliJPreferences;
 import saros.intellij.preferences.PropertiesComponentAdapter;
-import saros.intellij.project.filesystem.IntelliJWorkspaceImpl;
-import saros.intellij.project.filesystem.PathFactory;
 import saros.intellij.runtime.IntellijUISynchronizer;
 import saros.intellij.ui.eventhandler.ConnectingFailureHandler;
 import saros.intellij.ui.eventhandler.SessionStatusChangeHandler;
