+++ intellij/src/saros/intellij/context/SarosIntellijSessionContextFactory.java
@@ -7,8 +7,8 @@ import saros.intellij.editor.annotations.AnnotationManager;
 import saros.intellij.eventhandler.ApplicationEventHandlersFactory;
 import saros.intellij.eventhandler.ProjectEventHandlersFactory;
 import saros.intellij.eventhandler.project.ProjectClosedHandler;
+import saros.intellij.filesystem.SharedResourcesManager;
 import saros.intellij.followmode.FollowModeNotificationDispatcher;
-import saros.intellij.project.SharedResourcesManager;
 import saros.repackaged.picocontainer.MutablePicoContainer;
 import saros.session.ISarosSession;
 import saros.session.ISarosSessionContextFactory;
